requirements.txt
Django>=5.0,<5.1
sentence-transformers
jupyter
openai
pgvector
# psycopg2[binary]
psycopg[binary]
python-decouple
dj-database-url
dateparser
numpy
sqlalchemy 
llama-index
llama-index-llms-openai
llama-index-vector-stores-postgres